requirements.txt
Flask
Flask-Cors
requests
pandas
numpy
matplotlib
seaborn
plotly
tqdm
scikit-learn
PyJWT
openai
weaviate-client